Initial work on job control implementation
I decided to just use alists as a temporary way of having an associative
array for jobs and job ids. Which helped me to start writing some code
for job control with job ids. I added a bunch of code but most of
what I've added should be explained in comments.

More work on job control and the start of an api for domain job control

More work on job control, I've probably done as much as I can untill I
can ask some questions/get feedback.

Work on domain job control and the job control api

I've decided against using callbacks in the core job control
code, as I wasn't actually using them, I also added/finished a function
to wait for a job to finish, and added two new fields to the virJobObj
struct. one to keep track of the number of threads waiting for the job
to finish and one to define a maximum number of threads that can be
waiting at one time.