Replaced libxl job control functions with versions that use the new job control api
From as far as I can tell (on my machine at least) everything compiles
and passes all the tests, though I don't know if there are tests for job
control or not.

The next step is to replace the qemu job control functions, which will
be significantly more complicated, but the basic principals should be
similar to libxl.